Поправлена структура файлов
