+++ src/App.tsx
@@ -1,12 +1,13 @@
 import { Route, BrowserRouter, Routes } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
 import { AppRoute, AuthorizationStatus } from './const';
-import MainPage from './pages/Main';
-import OfferPage from './pages/Offer';
-import NotFoundPage from './pages/NotFound';
-import FavoritesPage from './pages/Favorites';
-import LoginPage from './pages/Login';
+import MainPage from './pages/MainPage';
+import OfferPage from './pages/OfferPage';
+import NotFoundPage from './pages/NotFoundPage';
+import FavoritesPage from './pages/FavoritesPage';
+import LoginPage from './pages/LoginPage';
 import PrivateRoute from './components/PrivateRoute';
+import Layout from './components/Layout';
 
 export type Place = {
   mark: string;
@@ -29,18 +30,23 @@ function App({ places }: AppProps): React.ReactNode {
     <HelmetProvider>
       <BrowserRouter>
         <Routes>
-          <Route path={AppRoute.Main} element={<MainPage places={places} />} />
-          <Route path={`${AppRoute.Offer}/:offerId`} element={<OfferPage />} />
-          <Route path={AppRoute.Login} element={<LoginPage />} />
-          <Route
-            path={AppRoute.Favorites}
-            element={
-              <PrivateRoute authorizationStatus={AuthorizationStatus.NoAuth}>
-                <FavoritesPage />
-              </PrivateRoute>
-            }
-          />
-          <Route path="*" element={<NotFoundPage />} />
+          <Route path={AppRoute.Main} element={<Layout />}>
+            <Route index element={<MainPage places={places} />} />
+            <Route
+              path={`${AppRoute.Offer}/:offerId`}
+              element={<OfferPage />}
+            />
+            <Route path={AppRoute.Login} element={<LoginPage />} />
+            <Route
+              path={AppRoute.Favorites}
+              element={
+                <PrivateRoute authorizationStatus={AuthorizationStatus.NoAuth}>
+                  <FavoritesPage />
+                </PrivateRoute>
+              }
+            />
+            <Route path="*" element={<NotFoundPage />} />
+          </Route>
         </Routes>
       </BrowserRouter>
     </HelmetProvider>
