+++ src/components/Favorites/index.tsx
@@ -1,0 +1,150 @@
+import { Helmet } from 'react-helmet-async';
+import { Link } from 'react-router-dom';
+import FavoritesPlaceCard from '../../components/FavoritesPlaceCard';
+import Footer from '../../components/Footer';
+
+//будет разбит на более мелкие компоненты позже
+function Favorites(): React.ReactNode {
+  return (
+    <>
+      <Helmet>
+        <title>6 cities. Favorites</title>
+      </Helmet>
+      <div className="page">
+        <main className="page__main page__main--favorites">
+          <div className="page__favorites-container container">
+            <section className="favorites">
+              <h1 className="favorites__title">Saved listing</h1>
+              <ul className="favorites__list">
+                <li className="favorites__locations-items">
+                  <div className="favorites__locations locations locations--current">
+                    <div className="locations__item">
+                      <Link className="locations__item-link" to="#">
+                        <span>Amsterdam</span>
+                      </Link>
+                    </div>
+                  </div>
+                  <div className="favorites__places">
+                    <FavoritesPlaceCard />
+
+                    <article className="favorites__card place-card">
+                      <div className="favorites__image-wrapper place-card__image-wrapper">
+                        <a href="#">
+                          <img
+                            className="place-card__image"
+                            src="img/room-small.jpg"
+                            width="150"
+                            height="110"
+                            alt="Place image"
+                          />
+                        </a>
+                      </div>
+                      <div className="favorites__card-info place-card__info">
+                        <div className="place-card__price-wrapper">
+                          <div className="place-card__price">
+                            <b className="place-card__price-value">&euro;80</b>
+                            <span className="place-card__price-text">
+                              &#47;&nbsp;night
+                            </span>
+                          </div>
+                          <button
+                            className="place-card__bookmark-button place-card__bookmark-button--active button"
+                            type="button"
+                          >
+                            <svg
+                              className="place-card__bookmark-icon"
+                              width="18"
+                              height="19"
+                            >
+                              <use xlinkHref="#icon-bookmark"></use>
+                            </svg>
+                            <span className="visually-hidden">
+                              In bookmarks
+                            </span>
+                          </button>
+                        </div>
+                        <div className="place-card__rating rating">
+                          <div className="place-card__stars rating__stars">
+                            <span style={{ width: '80%' }}></span>
+                            <span className="visually-hidden">Rating</span>
+                          </div>
+                        </div>
+                        <h2 className="place-card__name">
+                          <a href="#">Wood and stone place</a>
+                        </h2>
+                        <p className="place-card__type">Room</p>
+                      </div>
+                    </article>
+                  </div>
+                </li>
+
+                <li className="favorites__locations-items">
+                  <div className="favorites__locations locations locations--current">
+                    <div className="locations__item">
+                      <a className="locations__item-link" href="#">
+                        <span>Cologne</span>
+                      </a>
+                    </div>
+                  </div>
+                  <div className="favorites__places">
+                    <article className="favorites__card place-card">
+                      <div className="favorites__image-wrapper place-card__image-wrapper">
+                        <a href="#">
+                          <img
+                            className="place-card__image"
+                            src="img/apartment-small-04.jpg"
+                            width="150"
+                            height="110"
+                            alt="Place image"
+                          />
+                        </a>
+                      </div>
+                      <div className="favorites__card-info place-card__info">
+                        <div className="place-card__price-wrapper">
+                          <div className="place-card__price">
+                            <b className="place-card__price-value">&euro;180</b>
+                            <span className="place-card__price-text">
+                              &#47;&nbsp;night
+                            </span>
+                          </div>
+                          <button
+                            className="place-card__bookmark-button place-card__bookmark-button--active button"
+                            type="button"
+                          >
+                            <svg
+                              className="place-card__bookmark-icon"
+                              width="18"
+                              height="19"
+                            >
+                              <use xlinkHref="#icon-bookmark"></use>
+                            </svg>
+                            <span className="visually-hidden">
+                              In bookmarks
+                            </span>
+                          </button>
+                        </div>
+                        <div className="place-card__rating rating">
+                          <div className="place-card__stars rating__stars">
+                            <span style={{ width: '100%' }}></span>
+                            <span className="visually-hidden">Rating</span>
+                          </div>
+                        </div>
+                        <h2 className="place-card__name">
+                          <a href="#">White castle</a>
+                        </h2>
+                        <p className="place-card__type">Apartment</p>
+                      </div>
+                    </article>
+                  </div>
+                </li>
+              </ul>
+            </section>
+          </div>
+        </main>
+        <Footer />
+      </div>
+    </>
+  );
+}
+
+export default Favorites;
