+++ src/components/Layout/index.tsx
@@ -1,0 +1,13 @@
+import { Outlet } from 'react-router-dom';
+import Header from '../Header';
+
+export default function Layout() {
+  return (
+    <>
+      <Header />
+      <main>
+        <Outlet />
+      </main>
+    </>
+  );
+}
