+++ src/pages/FavoritesPage/index.ts
@@ -1,0 +1,2 @@
+import Favorites from '../../components/Favorites';
+export default Favorites;
