+++ src/pages/LoginPage/index.ts
@@ -1,0 +1,2 @@
+import Login from '../../components/Login';
+export default Login;
