+++ src/pages/NotFoundPage/index.ts
@@ -1,0 +1,2 @@
+import NotFound from '../../components/NotFound';
+export default NotFound;
