+++ src/pages/OfferPage/index.ts
@@ -1,0 +1,2 @@
+import Offer from '../../components/Offer';
+export default Offer;
